btd working for multiple connex components,  ternary constraints and preparing for nary and boosting  with variable elimination

+++ toulbar2/tb2store.hpp
@@ -122,7 +122,7 @@ public:
             }
             
             index = y - 1;
-            base = (int) pointers[y];
+            base = (long) pointers[y];
         }
     }
 };
